feet: 피드백리뷰 엔티티 Rating 타입 수정

+++ src/main/java/com/example/feedprep/domain/feedbackreview/dto/FeedbackReviewRequestDto.java
@@ -12,9 +12,9 @@ public class FeedbackReviewRequestDto {
 
 	private Long feedbackId;
 
-	@Min(value = 0)
-	@Max(value = 5)
-	private int rating;
+	@Min(value = 0L)
+	@Max(value = 5L)
+	private Long rating;
 
 	private String content;
 }
